refactor(service): simplifying weather_fetcher
